Refactor structure of GRPC speaker
