+++ services/src/grpc-speaker/grpc-api/src/main/java/org/openkilda/grpc/speaker/model/EnableLogMessagesResponse.java
@@ -17,26 +17,15 @@ package org.openkilda.grpc.speaker.model;
 
 import org.openkilda.messaging.model.grpc.OnOffState;
 
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.databind.PropertyNamingStrategy.SnakeCaseStrategy;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+import lombok.AllArgsConstructor;
 import lombok.Data;
-import lombok.EqualsAndHashCode;
-import lombok.NonNull;
 
 @Data
-@EqualsAndHashCode
-@JsonInclude(JsonInclude.Include.NON_NULL)
-public class LogOferrorsDto {
-    @NonNull
-    @JsonProperty("state")
-    private OnOffState state;
+@AllArgsConstructor
+@JsonNaming(SnakeCaseStrategy.class)
+public class EnableLogMessagesResponse {
 
-    @JsonCreator
-    public LogOferrorsDto(OnOffState state) {
-        if (state == null) {
-            throw new IllegalArgumentException("State must not be null.");
-        }
-        this.state = state;
-    }
+    private OnOffState enabled;
 }
