+++ services/src/grpc-speaker/grpc-api/src/main/java/org/openkilda/grpc/speaker/model/LicenseResponse.java
@@ -15,13 +15,15 @@
 
 package org.openkilda.grpc.speaker.model;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.databind.PropertyNamingStrategy.SnakeCaseStrategy;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
 @Data
 @AllArgsConstructor
-public class GrpcDeleteOperationResponse {
-    @JsonProperty("deleted")
-    private Boolean deleted;
+@JsonNaming(SnakeCaseStrategy.class)
+public class LicenseResponse {
+
+    private Boolean success;
 }
