+++ services/src/grpc-speaker/grpc-service/src/checkstyle/checkstyle-suppress-known-issues.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0"?>
+<!DOCTYPE suppressions PUBLIC
+        "-//Puppy Crawl//DTD Suppressions 1.2//EN"
+        "http://checkstyle.sourceforge.net/dtds/suppressions_1_2.dtd">
+
+<suppressions>
+    <suppress files="src/main/java/org/openkilda/grpc/speaker/GrpcSpeaker.java" checks="HideUtilityClassConstructor"/>
+</suppressions>
