+++ services/src/grpc-speaker/grpc-service/src/main/java/org/openkilda/grpc/speaker/config/MessageProducerConfig.java
