+++ services/src/grpc-speaker/grpc-service/src/main/java/org/openkilda/grpc/speaker/exception/GrpcRequestFailureException.java
@@ -1,4 +1,4 @@
-/* Copyright 2019 Telstra Open Source
+/* Copyright 2018 Telstra Open Source
  *
  *   Licensed under the Apache License, Version 2.0 (the "License");
  *   you may not use this file except in compliance with the License.
@@ -13,16 +13,26 @@
  *   limitations under the License.
  */
 
-package org.openkilda.grpc.speaker.model;
+package org.openkilda.grpc.speaker.exception;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.AllArgsConstructor;
-import lombok.Data;
+public class GrpcRequestFailureException extends RuntimeException {
 
-@Data
-@AllArgsConstructor
-public class LicenseResponse {
+    private final String message;
 
-    @JsonProperty("success")
-    private Boolean success;
+    private final Integer code;
+
+    public GrpcRequestFailureException(Integer code, String message) {
+        super(message);
+        this.message = message;
+        this.code = code;
+    }
+
+    @Override
+    public String getMessage() {
+        return message;
+    }
+
+    public Integer getCode() {
+        return code;
+    }
 }
