+++ services/src/grpc-speaker/grpc-service/src/main/java/org/openkilda/grpc/speaker/messaging/KafkaMessageProducer.java
