+++ services/src/grpc-speaker/grpc-service/src/main/resources/grpc-speaker.properties.example
