+++ services/src/grpc-speaker/grpc-service/src/main/resources/log4j2.xml.example
